refactor(forms): update structure
