+++ packages/ngx-kit/src/kit-forms/index.ts
@@ -1,8 +1,8 @@
-export * from './kit-forms.module';
-export * from './accessors/kit-input-value-accessor';
-export * from './accessors/middlewares/kit-limit-middleware';
-export * from './accessors/middlewares/kit-mask-middleware';
-export * from './accessors/middlewares/kit-math-parse-middleware';
 export * from './kit-check/kit-check.directive';
-export * from './math-parser';
+export * from './kit-input-value-accessor/kit-input-value-accessor.directive';
+export * from './middleware/kit-limit.middleware';
+export * from './middleware/kit-mask.middleware';
+export * from './middleware/kit-math-parse.middleware';
+export * from './middleware/math-parser';
+export * from './kit-forms.module';
 export * from './meta';
