+++ packages/ngx-kit/src/kit-forms/kit-forms.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { StylerModule } from '@ngx-kit/styler';
 import { KitCommonModule } from '../kit-common/kit-common.module';
-import { KitInputValueAccessor } from './accessors/kit-input-value-accessor';
+import { KitInputValueAccessorDirective } from './kit-input-value-accessor/kit-input-value-accessor.directive';
 import { KitCheckDirective } from './kit-check/kit-check.directive';
 
 @NgModule({
@@ -12,11 +12,11 @@ import { KitCheckDirective } from './kit-check/kit-check.directive';
     KitCommonModule,
   ],
   exports: [
-    KitInputValueAccessor,
+    KitInputValueAccessorDirective,
     KitCheckDirective,
   ],
   declarations: [
-    KitInputValueAccessor,
+    KitInputValueAccessorDirective,
     KitCheckDirective,
   ],
 })
