+++ packages/ngx-kit/src/kit-forms/kit-input-value-accessor/kit-input-value-accessor.directive.ts
@@ -5,7 +5,7 @@ import { KitInputMiddleware, kitInputMiddleware } from '../meta';
 
 export const KIT_INPUT_VALUE_ACCESSOR: any = {
   provide: NG_VALUE_ACCESSOR,
-  useExisting: forwardRef(() => KitInputValueAccessor),
+  useExisting: forwardRef(() => KitInputValueAccessorDirective),
   multi: true,
 };
 
@@ -23,7 +23,7 @@ export const KIT_INPUT_VALUE_ACCESSOR: any = {
   providers: [KIT_INPUT_VALUE_ACCESSOR],
 })
 // tslint:disable-next-line
-export class KitInputValueAccessor implements ControlValueAccessor {
+export class KitInputValueAccessorDirective implements ControlValueAccessor {
   private changes$ = new Subject<string>();
 
   private state: any;
